Refactor app runtime effect orchestration
